+++ apps/paper/src/Examples/Breathe/Breathe.tsx
@@ -3,12 +3,12 @@ import { StyleSheet, useWindowDimensions, View } from "react-native";
 import {
   BlurMask,
   vec,
-  Canvas2,
   Circle,
   Fill,
   Group,
   polar2Canvas,
   mix,
+  Canvas,
 } from "@shopify/react-native-skia";
 import type { SharedValue } from "react-native-reanimated";
 import { useDerivedValue } from "react-native-reanimated";
@@ -68,7 +68,7 @@ export const Breathe = () => {
 
   return (
     <View style={{ flex: 1 }}>
-      <Canvas2 style={styles.container} opaque>
+      <Canvas style={styles.container} opaque>
         <Fill color="rgb(36,43,56)" />
         <Group origin={center} transform={transform} blendMode="screen">
           <BlurMask style="solid" blur={40} />
@@ -76,7 +76,7 @@ export const Breathe = () => {
             return <Ring key={index} index={index} progress={progress} />;
           })}
         </Group>
-      </Canvas2>
+      </Canvas>
     </View>
   );
 };
