+++ apps/paper/src/Examples/Matrix/Matrix.tsx
@@ -1,6 +1,6 @@
 import {
   BlurMask,
-  Canvas2,
+  Canvas,
   Fill,
   Group,
   useClock,
@@ -41,7 +41,7 @@ export const Matrix = () => {
   }
   const symbols = font.getGlyphIDs("abcdefghijklmnopqrstuvwxyz");
   return (
-    <Canvas2 style={{ flex: 1 }} opaque>
+    <Canvas style={{ flex: 1 }} opaque>
       <Fill color="black" />
       <Group>
         <BlurMask blur={8} style="solid" />
@@ -60,6 +60,6 @@ export const Matrix = () => {
           ))
         )}
       </Group>
-    </Canvas2>
+    </Canvas>
   );
 };
