+++ packages/skia/src/renderer/Canvas.tsx
@@ -1,144 +1,63 @@
-import React, {
-  useEffect,
-  useCallback,
-  useMemo,
-  forwardRef,
-  useRef,
-} from "react";
-import type {
-  RefObject,
-  ReactNode,
-  MutableRefObject,
-  ForwardedRef,
-  FunctionComponent,
-} from "react";
-import type { LayoutChangeEvent } from "react-native";
+import { forwardRef, useEffect, useImperativeHandle, useMemo } from "react";
+import type { ViewProps } from "react-native";
 
-import { SkiaDomView } from "../views";
-import { Skia } from "../skia/Skia";
-import type { SkiaBaseViewProps } from "../views";
-import { SkiaJSDomView } from "../views/SkiaJSDomView";
+import { SkiaViewNativeId } from "../views/SkiaViewNativeId";
+import SkiaPictureViewNativeComponent from "../specs/SkiaPictureViewNativeComponent";
+import type { SkRect } from "../skia/types";
+import { SkiaSGRoot } from "../sksg/Reconciler";
+import { Skia } from "../skia";
 
-import { SkiaRoot } from "./Reconciler";
-import { NATIVE_DOM } from "./HostComponents";
+const NativeSkiaPictureView = SkiaPictureViewNativeComponent;
 
-export const useCanvasRef = () => useRef<SkiaDomView>(null);
-
-export interface CanvasProps extends SkiaBaseViewProps {
-  ref?: RefObject<SkiaDomView>;
-  children: ReactNode;
-  mode?: "default" | "continuous";
+interface CanvasProps extends ViewProps {
+  debug?: boolean;
+  opaque?: boolean;
 }
 
-const useOnSizeEvent = (
-  resultValue: SkiaBaseViewProps["onSize"],
-  onLayout?: (event: LayoutChangeEvent) => void
-) => {
-  return useCallback(
-    (event: LayoutChangeEvent) => {
-      if (onLayout) {
-        onLayout(event);
-      }
-      const { width, height } = event.nativeEvent.layout;
+export const Canvas = forwardRef(
+  ({ debug, opaque, children, ...viewProps }: CanvasProps, ref) => {
+    // Native ID
+    const nativeId = useMemo(() => {
+      return SkiaViewNativeId.current++;
+    }, []);
 
-      if (resultValue) {
-        resultValue.value = { width, height };
-      }
-    },
-    [onLayout, resultValue]
-  );
-};
+    // Root
+    const root = useMemo(() => new SkiaSGRoot(Skia, nativeId), [nativeId]);
 
-export const Canvas = forwardRef<SkiaDomView, CanvasProps>(
-  (
-    {
-      children,
-      style,
-      debug,
-      mode = "default",
-      onSize: _onSize,
-      onLayout: _onLayout,
-      ...props
-    },
-    forwardedRef
-  ) => {
-    const onLayout = useOnSizeEvent(_onSize, _onLayout);
-    const innerRef = useCanvasRef();
-    const ref = useCombinedRefs(forwardedRef, innerRef);
-    const redraw = useCallback(() => {
-      innerRef.current?.redraw();
-    }, [innerRef]);
-    const getNativeId = useCallback(() => {
-      const id = innerRef.current?.nativeId ?? -1;
-      return id;
-    }, [innerRef]);
-
-    const root = useMemo(
-      () => new SkiaRoot(Skia, NATIVE_DOM, redraw, getNativeId),
-      [redraw, getNativeId]
-    );
-
-    // Render effect
+    // Render effects
     useEffect(() => {
-      root!.render(children);
-    }, [children, root, redraw]);
+      root.render(children);
+    }, [children, root]);
 
     useEffect(() => {
       return () => {
-        root!.unmount();
+        root.unmount();
       };
     }, [root]);
-    if (NATIVE_DOM) {
-      return (
-        <SkiaDomView
-          ref={ref}
-          style={style}
-          root={root!.dom}
-          onLayout={onLayout}
-          debug={debug}
-          mode={mode}
-          {...props}
-        />
-      );
-    } else {
-      return (
-        <SkiaJSDomView
-          Skia={Skia}
-          mode={mode}
-          // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          ref={ref as any}
-          style={style}
-          root={root!.dom}
-          onLayout={onLayout}
-          debug={debug}
-          {...props}
-        />
-      );
-    }
-  }
-) as FunctionComponent<CanvasProps & React.RefAttributes<SkiaDomView>>;
 
-/**
- * Combines a list of refs into a single ref. This can be used to provide
- * both a forwarded ref and an internal ref keeping the same functionality
- * on both of the refs.
- * @param refs Array of refs to combine
- * @returns A single ref that can be used in a ref prop.
- */
-const useCombinedRefs = <T,>(
-  ...refs: Array<MutableRefObject<T> | ForwardedRef<T>>
-) => {
-  const targetRef = React.useRef<T>(null);
-  React.useEffect(() => {
-    refs.forEach((ref) => {
-      if (ref) {
-        if (typeof ref === "function") {
-          ref(targetRef.current);
-        } else {
-          ref.current = targetRef.current;
-        }
-      }
-    });
-  }, [refs]);
-  return targetRef;
-};
+    // Component methods
+    useImperativeHandle(ref, () => ({
+      makeImageSnapshot: (rect?: SkRect) => {
+        return SkiaViewApi.makeImageSnapshot(nativeId, rect);
+      },
+      makeImageSnapshotAsync: (rect?: SkRect) => {
+        return SkiaViewApi.makeImageSnapshot(nativeId, rect);
+      },
+      redraw: () => {
+        SkiaViewApi.requestRedraw(nativeId);
+      },
+      getNativeId: () => {
+        return nativeId;
+      },
+    }));
+    return (
+      <NativeSkiaPictureView
+        collapsable={false}
+        nativeID={`${nativeId}`}
+        debug={debug}
+        opaque={opaque}
+        {...viewProps}
+      />
+    );
+  }
+);
